+++ changelog.md
@@ -2,6 +2,7 @@
   - **Breaking Change** User can not directly change content via `tag.contnt = "abc"`
     - As content can be multiple lines, they are now a part of elements' sons  
 - add `add_as_first_son_of` to `ElementBase`
+- remove deprecated `add_as_son_of`
 
 ## 1.0.21
 - Safely write the file by writing to a temporary file first and then renaming it.